+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Update docs to explain SVG slide polyfill ([#117](https://github.com/marp-team/marpit/pull/117))
+
 ## v0.5.0 - 2018-12-28
 
 ### Added